+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/internal/ResourceDocument.java
@@ -66,6 +66,9 @@ import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 @Script(
 		name="typeRank", 
 		script="return params.ranks.getOrDefault(doc.resourceType.value, Integer.MAX_VALUE)")
+@Script(
+		name="snomedFirst", 
+		script="return (doc[\"toolingId\"].size() != 0 && doc.toolingId.value.equals(\"snomed\")) ? \"0\" : \"1\"")
 public final class ResourceDocument extends RevisionDocument {
 
 	public static final String TYPE = "resource";
@@ -211,7 +214,6 @@ public final class ResourceDocument extends RevisionDocument {
 		public static Expression upgradeOfs(Iterable<ResourceURI> upgradeOfs) {
 			return matchAny(Fields.UPGRADE_OF, Collections3.toImmutableSet(upgradeOfs).stream().map(ResourceURI::toString).collect(Collectors.toSet()));
 		}
-		
 	}
 
 	public static Builder builder() {